moved httpFetch function to root browser folder
